requirements.txt
paho-mqtt
python-dotenv
mysql-connector-python
